update calculateInterestPerSecond math

+++ src/interfaces/niftyapes/lending/ILending.sol
@@ -210,16 +210,16 @@ interface ILending is ILendingAdmin, ILendingEvents, ILendingStructs, IOffersStr
     /// @param interestRateBps The interest of the loan in bps
     /// @param duration The duration of the loan
     function calculateLenderInterestPerSecond(
-        uint128 amount,
-        uint96 interestRateBps,
-        uint32 duration
+        uint256 amount,
+        uint256 interestRateBps,
+        uint256 duration
     ) external pure returns (uint96);
 
     /// @notice Returns the protocolInterestRatePerSecond for a given set of terms
     ///         There is a set protocolInterestRateBps so no interestBps value is provided
     /// @param amount The amount of the loan
     /// @param duration The duration of the loan
-    function calculateProtocolInterestPerSecond(uint128 amount, uint32 duration)
+    function calculateProtocolInterestPerSecond(uint256 amount, uint256 duration)
         external
         view
         returns (uint96);
